pyproject.toml
[tool.poetry]
name = "gen-ai-demos"
version = "0.1.0"
description = ""
authors = ["Vicente Reig Rincon de Arellano <hey@vicente.services>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
dspy = "^2.5.43"
python-dotenv = "^1.0.1"
jupyter = "^1.1.1"
ipywidgets = "^8.1.5"
networkx = "^3.4.2"
matplotlib = "^3.10.0"
smolagents = "^1.9.2"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
